Move around assignments and expressions
