+++ test/unit/functions/validation/invalidFunctionAssignments.spec.ts
@@ -1,5 +1,5 @@
-import * as TSTLErrors from "../../../src/TSTLErrors";
-import * as util from "../../util";
+import * as TSTLErrors from "../../../../src/TSTLErrors";
+import * as util from "../../../util";
 import { invalidTestFunctionAssignments, invalidTestFunctionCasts } from "./functionPermutations";
 
 test.each(invalidTestFunctionAssignments)(
